+++ Puzzles/Util/Intcode/Instructions/BaseInstructions/BaseInstruction.cs
@@ -10,7 +10,7 @@ public abstract class BaseInstruction : IIntcodeInstruction
     public Parameter Parameter2 { get; set; }
     public Parameter Parameter3 { get; set; }
 
-    protected int ResultValue;
+    protected long ResultValue;
     
     protected BaseInstruction(Opcode opcode, (int m1, int m2, int m3) parameterModes)
     {
@@ -20,21 +20,21 @@ public abstract class BaseInstruction : IIntcodeInstruction
         Parameter3 = new Parameter(parameterModes.m3);
     }
 
-    public abstract void Run(int[] memory);
+    public abstract void Run(long[] memory);
 
     public abstract int Size();
 
-    public int Value()
+    public long Value()
     {
         return ResultValue;
     }
     
-    public int OutputAddress()
+    public long OutputAddress()
     {
         return Parameter3.Value;
     }
 
-    protected static int GetParameterValue(int[] memory, Parameter parameter)
+    protected static long GetParameterValue(long[] memory, Parameter parameter)
     {
         return parameter.ParameterMode == ParameterMode.Immediate ? parameter.Value : memory[parameter.Value];
     }
