+++ Puzzles/Util/Intcode/Instructions/EqualsInstruction.cs
@@ -8,7 +8,7 @@ public class EqualsInstruction : CompareInstruction
     public EqualsInstruction(Opcode opcode, (int m1, int m2, int m3) parameterModes) 
         : base(opcode, parameterModes) { }
     
-    public override void Run(int[] memory)
+    public override void Run(long[] memory)
     {
         ResultValue = GetParameterValue(memory, Parameter1) == GetParameterValue(memory, Parameter2) ? 1 : 0;
     }
